+++ vortex-array/src/compute/boolean.rs
@@ -1,167 +1,232 @@
-use vortex_error::{vortex_bail, VortexResult};
+use std::sync::Arc;
 
-use crate::array::BoolArray;
-use crate::{ArrayDType, ArrayData, IntoArrayVariant};
-pub trait AndFn {
-    /// Point-wise logical _and_ between two Boolean arrays.
-    ///
-    /// This method uses Arrow-style null propagation rather than the Kleene logic semantics.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use vortex_array::ArrayData;
-    /// use vortex_array::compute::and;
-    /// use vortex_array::IntoCanonical;
-    /// use vortex_array::accessor::ArrayAccessor;
-    /// let a = ArrayData::from_iter([Some(true), Some(true), Some(true), None, None, None, Some(false), Some(false), Some(false)]);
-    /// let b = ArrayData::from_iter([Some(true), None, Some(false), Some(true), None, Some(false), Some(true), None, Some(false)]);
-    /// let result = and(a, b)?.into_canonical()?.into_bool()?;
-    /// let result_vec = result.with_iterator(|it| it.map(|x| x.cloned()).collect::<Vec<_>>())?;
-    /// assert_eq!(result_vec, vec![Some(true), None, Some(false), None, None, None, Some(false), None, Some(false)]);
-    /// # use vortex_error::VortexError;
-    /// # Ok::<(), VortexError>(())
-    /// ```
-    fn and(&self, array: &ArrayData) -> VortexResult<ArrayData>;
+use arrow_array::cast::AsArray;
+use arrow_array::ArrayRef;
+use vortex_error::{vortex_bail, vortex_err, VortexError, VortexResult};
 
-    /// Point-wise Kleene logical _and_ between two Boolean arrays.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use vortex_array::ArrayData;
-    /// use vortex_array::compute::and_kleene;
-    /// use vortex_array::IntoCanonical;
-    /// use vortex_array::accessor::ArrayAccessor;
-    /// let a = ArrayData::from_iter([Some(true), Some(true), Some(true), None, None, None, Some(false), Some(false), Some(false)]);
-    /// let b = ArrayData::from_iter([Some(true), None, Some(false), Some(true), None, Some(false), Some(true), None, Some(false)]);
-    /// let result = and_kleene(a, b)?.into_canonical()?.into_bool()?;
-    /// let result_vec = result.with_iterator(|it| it.map(|x| x.cloned()).collect::<Vec<_>>())?;
-    /// assert_eq!(result_vec, vec![Some(true), None, Some(false), None, None, Some(false), Some(false), Some(false), Some(false)]);
-    /// # use vortex_error::VortexError;
-    /// # Ok::<(), VortexError>(())
-    /// ```
-    fn and_kleene(&self, array: &ArrayData) -> VortexResult<ArrayData>;
+use crate::arrow::FromArrowArray;
+use crate::encoding::Encoding;
+use crate::{ArrayDType, ArrayData, Canonical, IntoArrayVariant};
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum BinaryOperator {
+    And,
+    AndKleene,
+    Or,
+    OrKleene,
+    // AndNot,
+    // AndNotKleene,
+    // Xor,
 }
 
-pub trait OrFn {
-    /// Point-wise logical _or_ between two Boolean arrays.
-    ///
-    /// This method uses Arrow-style null propagation rather than the Kleene logic semantics.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use vortex_array::ArrayData;
-    /// use vortex_array::compute::or;
-    /// use vortex_array::IntoCanonical;
-    /// use vortex_array::accessor::ArrayAccessor;
-    /// let a = ArrayData::from_iter([Some(true), Some(true), Some(true), None, None, None, Some(false), Some(false), Some(false)]);
-    /// let b = ArrayData::from_iter([Some(true), None, Some(false), Some(true), None, Some(false), Some(true), None, Some(false)]);
-    /// let result = or(a, b)?.into_canonical()?.into_bool()?;
-    /// let result_vec = result.with_iterator(|it| it.map(|x| x.cloned()).collect::<Vec<_>>())?;
-    /// assert_eq!(result_vec, vec![Some(true), None, Some(true), None, None, None, Some(true), None, Some(false)]);
-    /// # use vortex_error::VortexError;
-    /// # Ok::<(), VortexError>(())
-    /// ```
-    fn or(&self, array: &ArrayData) -> VortexResult<ArrayData>;
-
-    /// Point-wise Kleene logical _or_ between two Boolean arrays.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use vortex_array::ArrayData;
-    /// use vortex_array::compute::or_kleene;
-    /// use vortex_array::IntoCanonical;
-    /// use vortex_array::accessor::ArrayAccessor;
-    /// let a = ArrayData::from_iter([Some(true), Some(true), Some(true), None, None, None, Some(false), Some(false), Some(false)]);
-    /// let b = ArrayData::from_iter([Some(true), None, Some(false), Some(true), None, Some(false), Some(true), None, Some(false)]);
-    /// let result = or_kleene(a, b)?.into_canonical()?.into_bool()?;
-    /// let result_vec = result.with_iterator(|it| it.map(|x| x.cloned()).collect::<Vec<_>>())?;
-    /// assert_eq!(result_vec, vec![Some(true), Some(true), Some(true), Some(true), None, None, Some(true), None, Some(false)]);
-    /// # use vortex_error::VortexError;
-    /// # Ok::<(), VortexError>(())
-    /// ```
-    fn or_kleene(&self, array: &ArrayData) -> VortexResult<ArrayData>;
+pub trait BinaryBooleanFn<Array> {
+    fn binary_boolean(
+        &self,
+        array: &Array,
+        other: &ArrayData,
+        op: BinaryOperator,
+    ) -> VortexResult<ArrayData>;
 }
 
-fn lift_boolean_operator<F, G>(
-    lhs: impl AsRef<ArrayData>,
-    rhs: impl AsRef<ArrayData>,
-    trait_fun: F,
-    bool_array_fun: G,
-) -> VortexResult<ArrayData>
+impl<E: Encoding + 'static> BinaryBooleanFn<ArrayData> for E
 where
-    F: Fn(&ArrayData, &ArrayData) -> Option<VortexResult<ArrayData>>,
-    G: FnOnce(BoolArray, &ArrayData) -> VortexResult<ArrayData>,
+    E: BinaryBooleanFn<E::Array>,
+    for<'a> &'a E::Array: TryFrom<&'a ArrayData, Error = VortexError>,
 {
-    let lhs = lhs.as_ref();
-    let rhs = rhs.as_ref();
-
-    if lhs.len() != rhs.len() {
-        vortex_bail!("Boolean operations aren't supported on arrays of different lengths")
+    fn binary_boolean(
+        &self,
+        lhs: &ArrayData,
+        rhs: &ArrayData,
+        op: BinaryOperator,
+    ) -> VortexResult<ArrayData> {
+        let array_ref = <&E::Array>::try_from(lhs)?;
+        let encoding = lhs
+            .encoding()
+            .as_any()
+            .downcast_ref::<E>()
+            .ok_or_else(|| vortex_err!("Mismatched encoding"))?;
+        BinaryBooleanFn::binary_boolean(encoding, array_ref, rhs, op)
     }
-
-    if !lhs.dtype().is_boolean() || !rhs.dtype().is_boolean() {
-        vortex_bail!("Boolean operations are only supported on boolean arrays")
-    }
-
-    if let Some(selection) = trait_fun(lhs, rhs) {
-        return selection;
-    }
-
-    if let Some(selection) = trait_fun(rhs, lhs) {
-        return selection;
-    }
-
-    // If neither side implements the trait, we try to expand the left-hand side into a `BoolArray`,
-    // which we know does implement it, and call into that implementation.
-    let lhs = lhs.clone().into_bool()?;
-
-    bool_array_fun(lhs, rhs)
 }
 
+pub trait AndFn<Array> {
+    fn and(&self, array: Array, other: &ArrayData) -> VortexResult<ArrayData>;
+}
+
+pub trait AndKleeneFn<Array> {
+    fn and_kleene(&self, array: &Array, other: &ArrayData) -> VortexResult<ArrayData>;
+}
+
+pub trait OrFn<Array> {
+    fn or(&self, array: &Array, other: &ArrayData) -> VortexResult<ArrayData>;
+}
+
+pub trait OrKleeneFn<Array> {
+    fn or_kleene(&self, array: &Array, other: &ArrayData) -> VortexResult<ArrayData>;
+}
+
+/// Point-wise logical _and_ between two Boolean arrays.
+///
+/// This method uses Arrow-style null propagation rather than the Kleene logic semantics.
+///
+/// # Examples
+///
+/// ```
+/// use vortex_array::ArrayData;
+/// use vortex_array::compute::and;
+/// use vortex_array::IntoCanonical;
+/// use vortex_array::accessor::ArrayAccessor;
+/// let a = ArrayData::from_iter([Some(true), Some(true), Some(true), None, None, None, Some(false), Some(false), Some(false)]);
+/// let b = ArrayData::from_iter([Some(true), None, Some(false), Some(true), None, Some(false), Some(true), None, Some(false)]);
+/// let result = and(a, b)?.into_canonical()?.into_bool()?;
+/// let result_vec = result.with_iterator(|it| it.map(|x| x.cloned()).collect::<Vec<_>>())?;
+/// assert_eq!(result_vec, vec![Some(true), None, Some(false), None, None, None, Some(false), None, Some(false)]);
+/// # use vortex_error::VortexError;
+/// # Ok::<(), VortexError>(())
+/// ```
 pub fn and(lhs: impl AsRef<ArrayData>, rhs: impl AsRef<ArrayData>) -> VortexResult<ArrayData> {
-    lift_boolean_operator(
-        lhs,
-        rhs,
-        |lhs, rhs| lhs.with_dyn(|lhs| lhs.and().map(|lhs| lhs.and(rhs))),
-        |lhs, rhs| lhs.and(rhs),
-    )
+    binary_boolean(lhs.as_ref(), rhs.as_ref(), BinaryOperator::And)
 }
 
+/// Point-wise Kleene logical _and_ between two Boolean arrays.
+///
+/// # Examples
+///
+/// ```
+/// use vortex_array::ArrayData;
+/// use vortex_array::compute::and_kleene;
+/// use vortex_array::IntoCanonical;
+/// use vortex_array::accessor::ArrayAccessor;
+/// let a = ArrayData::from_iter([Some(true), Some(true), Some(true), None, None, None, Some(false), Some(false), Some(false)]);
+/// let b = ArrayData::from_iter([Some(true), None, Some(false), Some(true), None, Some(false), Some(true), None, Some(false)]);
+/// let result = and_kleene(a, b)?.into_canonical()?.into_bool()?;
+/// let result_vec = result.with_iterator(|it| it.map(|x| x.cloned()).collect::<Vec<_>>())?;
+/// assert_eq!(result_vec, vec![Some(true), None, Some(false), None, None, Some(false), Some(false), Some(false), Some(false)]);
+/// # use vortex_error::VortexError;
+/// # Ok::<(), VortexError>(())
+/// ```
 pub fn and_kleene(
     lhs: impl AsRef<ArrayData>,
     rhs: impl AsRef<ArrayData>,
 ) -> VortexResult<ArrayData> {
-    lift_boolean_operator(
-        lhs,
-        rhs,
-        |lhs, rhs| lhs.with_dyn(|lhs| lhs.and_kleene().map(|lhs| lhs.and_kleene(rhs))),
-        |lhs, rhs| lhs.and_kleene(rhs),
-    )
+    binary_boolean(lhs.as_ref(), rhs.as_ref(), BinaryOperator::AndKleene)
 }
 
+/// Point-wise logical _or_ between two Boolean arrays.
+///
+/// This method uses Arrow-style null propagation rather than the Kleene logic semantics.
+///
+/// # Examples
+///
+/// ```
+/// use vortex_array::ArrayData;
+/// use vortex_array::compute::or;
+/// use vortex_array::IntoCanonical;
+/// use vortex_array::accessor::ArrayAccessor;
+/// let a = ArrayData::from_iter([Some(true), Some(true), Some(true), None, None, None, Some(false), Some(false), Some(false)]);
+/// let b = ArrayData::from_iter([Some(true), None, Some(false), Some(true), None, Some(false), Some(true), None, Some(false)]);
+/// let result = or(a, b)?.into_canonical()?.into_bool()?;
+/// let result_vec = result.with_iterator(|it| it.map(|x| x.cloned()).collect::<Vec<_>>())?;
+/// assert_eq!(result_vec, vec![Some(true), None, Some(true), None, None, None, Some(true), None, Some(false)]);
+/// # use vortex_error::VortexError;
+/// # Ok::<(), VortexError>(())
+/// ```
 pub fn or(lhs: impl AsRef<ArrayData>, rhs: impl AsRef<ArrayData>) -> VortexResult<ArrayData> {
-    lift_boolean_operator(
-        lhs,
-        rhs,
-        |lhs, rhs| lhs.with_dyn(|lhs| lhs.or().map(|lhs| lhs.or(rhs))),
-        |lhs, rhs| lhs.or(rhs),
-    )
+    binary_boolean(lhs.as_ref(), rhs.as_ref(), BinaryOperator::Or)
 }
 
+/// Point-wise Kleene logical _or_ between two Boolean arrays.
+///
+/// # Examples
+///
+/// ```
+/// use vortex_array::ArrayData;
+/// use vortex_array::compute::or_kleene;
+/// use vortex_array::IntoCanonical;
+/// use vortex_array::accessor::ArrayAccessor;
+/// let a = ArrayData::from_iter([Some(true), Some(true), Some(true), None, None, None, Some(false), Some(false), Some(false)]);
+/// let b = ArrayData::from_iter([Some(true), None, Some(false), Some(true), None, Some(false), Some(true), None, Some(false)]);
+/// let result = or_kleene(a, b)?.into_canonical()?.into_bool()?;
+/// let result_vec = result.with_iterator(|it| it.map(|x| x.cloned()).collect::<Vec<_>>())?;
+/// assert_eq!(result_vec, vec![Some(true), Some(true), Some(true), Some(true), None, None, Some(true), None, Some(false)]);
+/// # use vortex_error::VortexError;
+/// # Ok::<(), VortexError>(())
+/// ```
 pub fn or_kleene(
     lhs: impl AsRef<ArrayData>,
     rhs: impl AsRef<ArrayData>,
 ) -> VortexResult<ArrayData> {
-    lift_boolean_operator(
-        lhs,
-        rhs,
-        |lhs, rhs| lhs.with_dyn(|lhs| lhs.or_kleene().map(|lhs| lhs.or_kleene(rhs))),
-        |lhs, rhs| lhs.or_kleene(rhs),
-    )
+    binary_boolean(lhs.as_ref(), rhs.as_ref(), BinaryOperator::OrKleene)
+}
+
+fn binary_boolean(lhs: &ArrayData, rhs: &ArrayData, op: BinaryOperator) -> VortexResult<ArrayData> {
+    if lhs.len() != rhs.len() {
+        vortex_bail!("Boolean operations aren't supported on arrays of different lengths")
+    }
+    if !lhs.dtype().is_boolean() || !rhs.dtype().is_boolean() {
+        vortex_bail!("Boolean operations are only supported on boolean arrays")
+    }
+
+    // If LHS is constant, then we make sure it's on the RHS.
+    if lhs.is_constant() && !rhs.is_constant() {
+        return binary_boolean(rhs, lhs, op);
+    }
+
+    // Check if either LHS or RHS supports the operation directly.
+    if let Some(f) = lhs.encoding().binary_boolean_fn(lhs, rhs) {
+        return f.binary_boolean(lhs, rhs, op);
+    } else {
+        log::debug!(
+            "No boolean implementation found for LHS {}, RHS {}, and operator {:?}",
+            lhs.encoding().id(),
+            rhs.encoding().id(),
+            op,
+        );
+    }
+    if let Some(f) = rhs.encoding().binary_boolean_fn(rhs, lhs) {
+        return f.binary_boolean(rhs, lhs, op);
+    } else {
+        log::debug!(
+            "No boolean implementation found for LHS {}, RHS {}, and operator {:?}",
+            rhs.encoding().id(),
+            lhs.encoding().id(),
+            op,
+        );
+    }
+
+    // If neither side implements the trait, we try to expand the left-hand side into a `BoolArray`,
+    // which we know does implement it, and call into that implementation.
+    arrow_boolean(lhs.clone(), rhs.clone(), op)
+}
+
+/// Implementation of `BinaryBooleanFn` using the Arrow crate.
+///
+/// Note that other encodings should handle a constant RHS value, so we can assume here that
+/// the RHS is not constant and expand to a full array.
+pub(crate) fn arrow_boolean(
+    lhs: ArrayData,
+    rhs: ArrayData,
+    operator: BinaryOperator,
+) -> VortexResult<ArrayData> {
+    let nullable = lhs.dtype().is_nullable() || rhs.dtype().is_nullable();
+
+    let lhs = Canonical::Bool(lhs.into_bool()?)
+        .into_arrow()?
+        .as_boolean()
+        .clone();
+    let rhs = Canonical::Bool(rhs.into_bool()?)
+        .into_arrow()?
+        .as_boolean()
+        .clone();
+
+    let array = match operator {
+        BinaryOperator::And => arrow_arith::boolean::and(&lhs, &rhs)?,
+        BinaryOperator::AndKleene => arrow_arith::boolean::and_kleene(&lhs, &rhs)?,
+        BinaryOperator::Or => arrow_arith::boolean::or(&lhs, &rhs)?,
+        BinaryOperator::OrKleene => arrow_arith::boolean::or_kleene(&lhs, &rhs)?,
+    };
+
+    Ok(ArrayData::from_arrow(Arc::new(array) as ArrayRef, nullable))
 }
 
 #[cfg(test)]
